yt2mp3/requirements.txt
yt_dlp==2023.1.6
google-cloud-storage==2.7.0
